test: fold the /mcp dual-era tests into mcp.test.ts
A second mock/restore cycle of ./supabase.js in its own file broke
middleware.test.ts's mock on Linux CI (the third mock.module of the same path
no longer reached the already-loaded consumer); one mock window per process.
